pyproject.toml
[project]
name = "videobgremover"
version = "0.1.1"
description = "Official Python SDK for VideoBGRemover API - Remove video backgrounds with AI and compose videos with FFmpeg"
readme = "README.md"
authors = [
    { name = "Paul", email = "paul@videobgremover.com" }
]
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.11.9",
    "requests>=2.32.5",
]

[project.urls]
Homepage = "https://videobgremover.com"
Documentation = "https://docs.videobgremover.com"
Repository = "https://github.com/videobgremover/videobgremover-python"
Issues = "https://github.com/videobgremover/videobgremover-python/issues"

[project.optional-dependencies]
dev = [
    "build>=1.3.0",
    "mypy>=1.18.2", 
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.13.1",
    "twine>=6.2.0",
    "responses>=0.24.0",
    "python-dotenv>=1.0.0",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "build>=1.3.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.1.1",
    "responses>=0.25.8",
    "ruff>=0.13.1",
    "twine>=6.2.0",
    "types-requests>=2.31.0",
]
